test: re-vendor migrate/integration (stacked) lib into VM charm
Bumps the vendored postgresql-charms-single-kernel from 16.2.6 to
16.2.7, reflecting the migrate/integration branch of the library which
adds integration-test helpers. All 237 unit tests continue to pass
with no charm source changes.
